improve simulation of the github merge button
Now, we check out the target, add the fork as a remote,
fetch it, and merge the SHA under test. Added a helper
function that does all this in the `task` module; tests
should detect whether they are doing a straightforward
clone or a PR merge based on the presence of a base
repo URL and ref. See `build` for an example.